etcdctl installation: Download for correct system architecture (#40)
TL;DR Script now works on Ubuntu 20.04 VM on MacBook Pro with M1

The script worked very nicely in a Ubuntu VM on a MacBook Pro with M1. That is, until I made to section 16 in the CKS course. I noticed that etcdctl on my node 'controlplane' was for amd64 (-bash: /usr/bin/etcdctl: cannot execute binary file: Exec format error). Albeit a little outside the scope for the course (as most folks will probably not use a set of local VMs), this change now accounts for the build architecture of the master node. Thank you for the excellent CKS course!